Allow PL78 to build with all corrections from BCU tape.
Problem:
modset ST5A362 adds a new capsule named S$CMGFS but does not
add the correct set of control cards to the PL78I deck to correctly
build the new capsule. This results in a "BAD STR5LIB" error.

Solution:
We add the correct control cards to PL78I. This consisists of
deleting the RETURN(BASE) control card because a WEOF is added to deck
SKELETON, which causes everything for the added capsule to end up past
EOF on BASE. Next we insert the correct loader and CAPSULE statements to
build the capsule properly and have it end up on the library with
everything else.  I eventually used the SORT5 deck on NOS2.8.7 to help
me come up with the right sequence of control cards. SORT5 at level 871
also contains correction ST5A362.

Notes:
COPYLM complains about S$ISQRT not being found. this is a warning
message. ST5A362 removes a reference to this unused function from
deck SKELETON, and since it's still a deck it gets assembled, but
there's nothing to merge it with so it's discarded.

See common deck *HELP and common deck SKELETON plus a listing of SKELGEN
for the full gory details of what's going on.  this is a pattern also
used by the QU3 and CDCS products.